move signing to seperate module
